fix(driver/modern_bpf): avoid type confusion setting snaplen

+++ driver/modern_bpf/programs/tail_called/events/syscall_dispatched_events/pwrite64.bpf.c
@@ -70,10 +70,10 @@ int BPF_PROG(pwrite64_x,
 	/* If the syscall doesn't fail we use the return value as `size`
 	 * otherwise we need to rely on the syscall parameter provided by the user.
 	 */
-	uint16_t bytes_to_read = ret > 0 ? ret : extract__syscall_argument(regs, 2);
+	int64_t bytes_to_read = ret > 0 ? ret : extract__syscall_argument(regs, 2);
 	uint16_t snaplen = maps__get_snaplen();
 	apply_dynamic_snaplen(regs, &snaplen, false);
-	if(snaplen > bytes_to_read)
+	if((int64_t)snaplen > bytes_to_read)
 	{
 		snaplen = bytes_to_read;
 	}
